+++ mempool/mock/mempool.go
@@ -1,0 +1,36 @@
+package mock
+
+import (
+	abci "github.com/tendermint/tendermint/abci/types"
+	mempl "github.com/tendermint/tendermint/mempool"
+	"github.com/tendermint/tendermint/types"
+)
+
+// Mempool is an empty implementation of a Mempool, useful for testing.
+type Mempool struct{}
+
+var _ mempl.Mempool = Mempool{}
+
+func (Mempool) Lock()     {}
+func (Mempool) Unlock()   {}
+func (Mempool) Size() int { return 0 }
+func (Mempool) CheckTx(_ types.Tx, _ func(*abci.Response)) error {
+	return nil
+}
+func (Mempool) CheckTxWithInfo(_ types.Tx, _ func(*abci.Response),
+	_ mempl.TxInfo) error {
+	return nil
+}
+func (Mempool) ReapMaxBytesMaxGas(_, _ int64) types.Txs { return types.Txs{} }
+func (Mempool) Update(
+	_ int64,
+	_ types.Txs,
+	_ mempl.PreCheckFunc,
+	_ mempl.PostCheckFunc,
+) error {
+	return nil
+}
+func (Mempool) Flush()                        {}
+func (Mempool) FlushAppConn() error           { return nil }
+func (Mempool) TxsAvailable() <-chan struct{} { return make(chan struct{}) }
+func (Mempool) EnableTxsAvailable()           {}
